Promote analysis results deprecation (#1532)
* Replace `deprecate_arg` with `warn_from_qe` for issuing deprecation
warnings from internal qiskit-experiments code. `deprecate_arg` was not
working for the cases it was being used for (deprecating the default
value of a keyword argument, deprecating the type of an argument in a
function that could be called at different stack levels). `warn_from_qe`
blames the caller of qiskit-experiments. The docstring deprecation
message has to be added separately.
* Promote the deprecation of
`ExperimentData.add_analysis_results(results)`. This was previously a
pending deprecation. Now results must be added via keyword arguments for
their properties. All internal usage was updated not to issue
deprecation warnings. `add_analysis_result` was added as an alias to
reflect that only one result should be added at a time now.
* Promote deprecation of
`ExperimentData.analysis_results(dataframe=False)`. All internal usage
was updated not to trigger deprecation warnings. The default remains
`dataframe=False` for backwards compatibility. This default should not
change in the near term because this is a major user-facing change.
* Update `RBAnalysis` to accept the dataframe version of the analysis
result as the `epg_1_qubit` option.
* Add `AnalysisResultData.as_table_element` for convenience in passing a
result data object to `add_analysis_result`.
* Use `warn_from_qe` to warn about slice and integer arguments to
`analysis_results`. This might result in warnings being visible that
were previously hidden by stack level.