Fix cases of passing `basis_gates` or `coupling_map` to `transpile` (#1540)
In Qiskit 2.0, passing a `backend` and `basis_gates` /`coupling_map` to
`transpile` triggers a warning about this possible producing
inconsistent results. Two cases of this warning are addressed here:

* The `basis_gates` transpile option of `FineFrequency` was removed.
Transpilation should use the gates in the backend and not transpile to a
predetermined set that might not match the backend. Using a
predetermined set worked okay in the past when this experiment was used
with features that would attach pulse gate calibrations for the basis
gates to the circuits but those features were removed in 0.9.0.
* A filter previously added to ignore this warning for one test was
accidentally committed as "error" instead of "ignore" and this was not
noticed because the warning only occurs for Qiskit 2.0. The filter is
switched to "ignore" here. The test needs to pass different coupling
maps to experiments so it has to trigger the warning.