Ensure docs option parsing does not drop final argument (#1545)
The custom options parser for experiment and analysis class
documentation loops through the options methods' docstrings to pull out
the documentation for the different options. This loop parsed the
docstring into options by accumulating an option across multiple lines
and then adding to the parsed set when it hit a less indented line.
Starting with Python 3.13, the trailing whitespace of docstrings is
stripped automatically. So the final newline that docstrings usually had
is no longer present and a less indented line is never hit for the final
option to be parsed, resulting in that option being dropped from the
documentation.

Here the loop was reworked to accumulate the options more eagerly so
that the final option is not dropped.