Add measure to test circuit to avoid warnings (#1544)
Recently, qiskit-ibm-runtime's `SamplerV2.run` started warning about
circuits that do not have any measurements. Here a measurement is added
to the circuit of `FakeExperiment` so that all the tests using it do not
issue warnings.

Additionally, the number of qubits set on the test backend was corrected
in a couple tests. The incorrect qubit number was exposed by transpiler
errors after adding the `measure_all()` call that added measurements on
higher index qubits than the backend was set to support.