Update PR template to match expectations (#1547)
This change has two goals:

* More prominently emphasize that the PR description becomes the commit
message so the template content needs to be deleted.
* Deemphasize making the PR description all encompassing. The previous
version of the template had summary and details sections and encouraged
talking through alternative solutions considered and how the change
affects users. That is a bit much to end up in the final commit message
and is better pushed to PR comments or the release notes.