Address Qiskit 2.0 linting issues (#1551)
Some parts of the code that retain deprecated features for Qiskit 1.0
are causing pylint errors now that Qiskit 2.0 has been released and is
being used in the pylint environment. Here explicit disabling of pylint
in those areas is added and can remain until those parts of the code can
be removed.