Ignore warning about std_dev==0 in ufloat_params (#1556)
As of Uncertainties 3.2.3, `ufloat()` warns when passed a standard
deviation of 0 because that can lead to problems (see the discussion
[here](https://github.com/lmfit/uncertainties/issues/283).  The
recommendation is to use a plain float instead but we want
`ufloat_params` to return all `UFloats` so that `.n` and `.s` always
exist and the user doesn't need to handle type checking on curve data
results, so here we just ignore the warning.

A standard deviation of 0 is usually not a problem. Uncertainties
propagates errors for all of its variables even when they have 0
standard deviation. If the propagation of one variable leads to a `nan`
or `inf` value, the standard deviation will take on that value even
though, since the variable's standard deviation was 0, the error
propagation could have been skipped for that variable.

(cherry picked from commit 29155d054457fd90f3fcca0c4512518311382da7)