Address deprecated use of Qiskit's QuantumVolume (#1565)
The `QuantumVolume` class in Qiskit has been deprecated in version 2.2
in favor of the `quantum_volume` function. Here the latter is swapped
out for the former. In doing so, a bug with `quantum_volume` was
discovered and worked around. `quantum_volume` was added in Qiskit 1.3,
so the minimum version of Qiskit was raised from 0.45 to 1.3 since that
is a reasonably old version rather than trying to keep using
`QuantumVolume` with older versions of Qiskit.