Combine stable and release docs workflows (#1570)
Currently, these workflows conflict because they both trigger off a tag
and try to push the 'gh-pages' branch. Whichever workflow finishes
second always fails. Here we change to push to both places from the
stable workflow but only push to the top level directory that the
release workflow was pushing when the tag appears to be the latest
release.

Closes #1554