Prepare 0.11.0 release (#1571)
Following the steps in `CONTRIBUTING.md`, this change moves the old
release notes into the main document and fixes up the remaining release
notes for the new release. The tense and voice of the different notes
were unified and some notes were moved to different categories. It is
hard to say whether updating something not to use an API that was just
deprecated should be a bug fix or an upgrade note. Here the notes lean
toward upgrade notes with only a change that affects existing behavior
(slow performance with old qiskit-ibm-runtime) counting as a bug fix.