Suppress new scs warning from docs output (#1580)
jupyter-sphinx causes the docs build to fail for any warning so we need
to suppress warnings for the docs to build. I am not sure if there is
something we should be doing in this specific warning's case. The
warning text is:

> `/.tox/docs/lib/python3.13/site-packages/scs/__init__.py:83:
UserWarning: `
> `Converting A to a CSC (compressed sparse column) matrix; may take a
while.`

The code emitting the warning in scs has not changed recently but
something else in the pipeline must have changed to generate a matrix
differently than it did before.

(cherry picked from commit d76642b96ef6404095e8622f80463a22ad7bb1bc)