Work around lmfit limiting RBAnalysis to 10,000 Cliffords (#1595)
lmfit (through its asteval dependnecy) limits exponentiation in its
expression models to 10,000. The limit is there to avoid denial of
service attacks where user input could be substituted into a model and
Python integers which are unbounded and could take up exponential
amounts of memory. In qiskit-experiments, this case is not a concern.
The workaround implemented here is to divide the exponential in the
curve by 10 and then exponentiate the result by 10. This change allows
up to 100,000 Cliffords which should be enough. The alternative
workaround is to change `asteval.astutils.MAX_EXPONENT` to a
sufficiently large value.

Closes
https://github.com/qiskit-community/qiskit-experiments/issues/1594

(cherry picked from commit 643c20ac50ea49078960927fff0a683058c4f427)